Added functions to process and fill packets, but realised that I will need a constant memory slice and wondered if that was possible. Turns out it _is_ possible, I guess it is C after all, but I forgot that different implementations will have different byte orders; in this case, my host machine and the target mcu are both LSB. Now I can create a memory pool to build from.

+++ app/inc/dringbuffer.h
@@ -9,15 +9,15 @@
 
 /*********** ringbuffer  *************/
 typedef struct ring_buf_t {
-    uint16_t* buffer;    // pointer to buffer
+    uint8_t* buffer;    // pointer to buffer
     uint32_t head;      // head index (write)
     uint32_t tail;      // tail index (read)
     uint32_t mask;      // HAS to be (2^n) - 1
 } ring_buf_t;
 
-extern error_t dring_buf_setup(ring_buf_t* rb, uint16_t* buffer, uint32_t size);
+extern error_t dring_buf_setup(ring_buf_t* rb, uint8_t* buffer, uint32_t size);
 extern int dring_buf_empty(ring_buf_t* rb);
-extern error_t dring_buf_write(ring_buf_t* rb, uint16_t byte);
-extern error_t dring_buf_read(ring_buf_t* rb, uint16_t* byte);
+extern error_t dring_buf_write(ring_buf_t* rb, uint8_t byte);
+extern error_t dring_buf_read(ring_buf_t* rb, uint8_t* byte);
 
 #endif // DRING_BUF_H
